backend/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
pydantic==2.13.3
python-dotenv==1.2.2
langchain==1.2.17
langchain-core==1.3.2
langchain-openai==1.2.1
langgraph==1.1.10
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.13
cerebras_cloud_sdk==1.67.0
google_search_results==2.4.2
requests==2.33.1
httpx==0.28.1
python-multipart==0.0.28
langsmith==0.8.0
deepeval>=2.0.0
pytest>=8.0.0
